Split frontend behavior diagnostic tests
